refactor: 멤버 픽스쳐 분리
