+++ backend/sokdak/src/test/java/com/wooteco/sokdak/auth/service/AuthServiceTest.java
@@ -1,12 +1,11 @@
 package com.wooteco.sokdak.auth.service;
 
+import static com.wooteco.sokdak.util.fixture.MemberFixture.INVALID_LOGIN_REQUEST;
+import static com.wooteco.sokdak.util.fixture.MemberFixture.VALID_LOGIN_REQUEST;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
-import com.wooteco.sokdak.auth.SignupHelper;
 import com.wooteco.sokdak.auth.dto.AuthInfo;
-import com.wooteco.sokdak.auth.dto.LoginRequest;
 import com.wooteco.sokdak.auth.exception.LoginFailedException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -24,10 +23,7 @@ class AuthServiceTest {
     @DisplayName("로그인 기능")
     @Test
     void login() {
-        SignupHelper.signUp();
-        LoginRequest loginRequest = new LoginRequest(SignupHelper.USERNAME, SignupHelper.PASSWORD);
-
-        AuthInfo authInfo = authService.login(loginRequest);
+        AuthInfo authInfo = authService.login(VALID_LOGIN_REQUEST);
 
         assertThat(authInfo.getId()).isNotNull();
     }
@@ -35,9 +31,7 @@ class AuthServiceTest {
     @DisplayName("존재하지 않는 회원정보로 로그인 시 예외 발생")
     @Test
     void login_Exception() {
-        LoginRequest loginRequest = new LoginRequest(SignupHelper.USERNAME, SignupHelper.PASSWORD);
-
-        assertThatThrownBy(() -> authService.login(loginRequest))
+        assertThatThrownBy(() -> authService.login(INVALID_LOGIN_REQUEST))
                 .isInstanceOf(LoginFailedException.class);
     }
 }
