+++ backend/sokdak/src/test/java/com/wooteco/sokdak/post/controller/PostControllerTest.java
@@ -1,8 +1,8 @@
 package com.wooteco.sokdak.post.controller;
 
-import static com.wooteco.sokdak.util.HttpMethodFixture.getExceptionMessage;
-import static com.wooteco.sokdak.util.HttpMethodFixture.httpGet;
-import static com.wooteco.sokdak.util.HttpMethodFixture.httpPost;
+import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.getExceptionMessage;
+import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpGet;
+import static com.wooteco.sokdak.util.fixture.HttpMethodFixture.httpPost;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
