+++ backend/sokdak/src/test/java/com/wooteco/sokdak/util/fixture/HttpMethodFixture.java
@@ -1,4 +1,4 @@
-package com.wooteco.sokdak.util;
+package com.wooteco.sokdak.util.fixture;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
