+++ backend/sokdak/src/test/resources/import.sql
@@ -1,0 +1,1 @@
+insert into member (username, nickname, password) values ('chris', 'chrisNickname', '6297d64078fc9abcfe37d0e2c910d4798bb4c04502d7dd1207f558860c2b382e')
